hw4.1
